Further concurrency changes
Had another look over the code and replaced some instances of particularly heavy tasks with actual threads (it's fine to use blocking calls inside those). In general, whenever there is a pptx::task created in the code, it should only contain .get() or .wait() calls that are pre-guaranteed to complete instantly (by accepting a task as the continuation parameter).